+++ packages/server/shared/src/lib/system/system.ts
@@ -90,8 +90,7 @@ const systemPropDefaultValues: Partial<Record<SystemProp, string>> = {
   [AppSystemProp.DOCS_MCP_SERVER_PATH]: '/root/.mcp/docs.openops.com',
   [AppSystemProp.LOAD_EXPERIMENTAL_MCP_TOOLS]: 'false',
   [SharedSystemProp.AWS_ENABLE_IMPLICIT_ROLE]: 'false',
-  [AppSystemProp.OPENOPS_MCP_SERVER_PATH]:
-    '/Users/leylabuechel/Documents/openops-mcp',
+  [AppSystemProp.OPENOPS_MCP_SERVER_PATH]: '/root/.mcp/openops-mcp',
   [AppSystemProp.SAMPLE_DATA_SIZE_LIMIT_KB]: '5',
 };
